requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
attrs==25.3.0
automated-interpretability==0.0.13
babe==0.0.7
beartype==0.14.1
better-abc==0.0.3
blobfile==2.1.1
boostedblob==0.15.6
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
config2py==0.1.37
contourpy==1.3.3
cycler==0.12.1
datasets==3.6.0
dill==0.3.8
docstring_parser==0.17.0
dol==0.3.19
dotenv==0.9.9
einops==0.8.1
fancy-einsum==0.0.3
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.45
gprof2dot==2025.4.14
graze==0.1.29
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.34.4
hydra-core==1.3.2
i2==0.1.48
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
jaxtyping==0.3.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lxml==4.9.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
networkx==3.5
nltk==3.9.1
numpy==1.26.4
omegaconf==2.3.0
orjson==3.11.1
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pillow==11.3.0
platformdirs==4.3.8
plotly==5.24.1
plotly-express==0.4.1
pluggy==1.6.0
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
py2store==0.1.20
pyarrow==21.0.0
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
pytest-profiling==1.8.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.0.0
regex==2025.7.34
requests==2.32.4
rich==14.1.0
sae-lens==6.5.2
safetensors==0.4.5
scikit-learn==1.7.1
scipy==1.16.1
sentencepiece==0.2.0
sentry-sdk==2.34.1
shellingham==1.5.4
simple-parsing==0.1.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
statsmodels==0.14.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
traitlets==5.14.3
transformer-lens==2.16.1
transformers==4.55.0
transformers-stream-generator==0.0.5
typeguard==4.4.4
typer==0.12.5
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvloop==0.21.0
wadler_lindig==0.1.7
wandb==0.21.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.22.0
